Refactor: auth 인증/인가 프로젝트 구조 변경[박한솔]
