+++ web3-credential-server/src/main/java/web3/auth/dto/JwtResponse.java
@@ -1,4 +1,4 @@
-package web3.util;
+package web3.auth.dto;
 
 public class JwtResponse {
     private String accessToken;
