+++ web3-credential-server/src/main/resources/application.properties
@@ -40,7 +40,7 @@ jwt.refresh-token-validity=86400000
 #spring.main.allow-bean-definition-overriding=true
 
 #logging.level.org.springframework.web: DEBUG
-#logging.level.web3.util.JwtTokenUtil: DEBUG
+#logging.level.web3.auth.util.JwtTokenUtil: DEBUG
 
 #???? ??
 #kakao.client-id=
